simple_gan/requirements.txt
pyyaml

torch
torchvision
tensorboard
